+++ src/audio/source.rs
@@ -1,7 +1,7 @@
 use super::effect::Effect;
 
 pub trait Source: Send {
-    fn output_audio(&mut self, sample: usize, data: &mut [f32]);
+    fn output_audio(&mut self, sample: usize, data: &mut [f64]);
 }
 
 pub struct FxSource {
@@ -16,7 +16,7 @@ impl FxSource {
 }
 
 impl Source for FxSource {
-    fn output_audio(&mut self, sample: usize, data: &mut [f32]) {
+    fn output_audio(&mut self, sample: usize, data: &mut [f64]) {
         self.source.output_audio(sample, data);
         for effect in self.effects.iter_mut() {
             effect.process_audio(sample, data);
